Major CSS chagnes
Removed inline styling for the drop down navigation in the HTML file, and commented out the original CSS file to start a new CSS file. I also moved the calculator screen input to be with in the calculator keys so it works with the grid layout.

Started a new CSS file so that I can rework the CSS from ground up.